requirements.txt
Flask
cryptography
steganography
